requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-pytest==0.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
urllib3==2.0.4
